rating prediction: cast

+++ src/RatingPrediction/RatingPrediction.cs
@@ -267,7 +267,7 @@ class RatingPrediction
 			RatingEval.DisplayResults(RatingEval.Evaluate(recommender, test_data));
 			Console.WriteLine(" " + iterative_recommender.NumIter);
 
-			for (int i = iterative_recommender.NumIter + 1; i <= max_iter; i++)
+			for (int i = (int) iterative_recommender.NumIter + 1; i <= max_iter; i++)
 			{
 				TimeSpan time = Utils.MeasureTime(delegate() {
 					iterative_recommender.Iterate();
